services/classify_worker/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-pubsub
google-cloud-firestore
pydantic
